refactor : DibsRepository 에 QueryDSL 적용

+++ src/main/java/com/example/sparta/domain/dibs/service/DibsServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.sparta.domain.dibs.service;
 
 import com.example.sparta.domain.dibs.dto.DibsResponseDto;
 import com.example.sparta.domain.dibs.entity.Dibs;
+import com.example.sparta.domain.dibs.repository.DibsQueryRepository;
 import com.example.sparta.domain.dibs.repository.DibsRepository;
 import com.example.sparta.domain.store.entity.Store;
 import com.example.sparta.domain.store.repository.StoreRepository;
@@ -16,13 +17,14 @@ public class DibsServiceImpl implements DibsService {
 
   private final StoreRepository storeRepository;
   private final DibsRepository dibsRepository;
+  private final DibsQueryRepository dibsQueryRepository;
 
   @Override
   public DibsResponseDto createDibs(Long id, User user) {
     Store store = storeRepository.findById(id)
         .orElseThrow(() -> new NoSuchElementException("해당 가계는 찾을수 없어요"));
 
-    if (dibsRepository.findDibsByStoreAndUser(store, user).isPresent()) {
+    if (dibsQueryRepository.findDibsByStoreAndUser(store, user) != null) {
       throw new IllegalArgumentException("이미 찜한 가게입니다.");
     }
 
@@ -41,8 +43,10 @@ public class DibsServiceImpl implements DibsService {
         .orElseThrow(() -> new NoSuchElementException("해당 가계는 찾을수 없어요"));
 
     try {
-      Dibs dibs = dibsRepository.findDibsByStoreAndUser(store, user)
-          .orElseThrow(() -> new NoSuchElementException("해당 찜을 찾을수 없어요"));
+      Dibs dibs = dibsQueryRepository.findDibsByStoreAndUser(store, user);
+      if (dibs == null) {
+        throw new NoSuchElementException("해당 찜을 찾을수 없어요");
+      }
       dibsRepository.delete(dibs);
       return storeId;
     } catch (Exception e) {
